COMMERCE-11740 autogenerated code (gw buildService)

+++ modules/apps/commerce/commerce-inventory-api/src/main/java/com/liferay/commerce/inventory/model/CommerceInventoryWarehouseItemModel.java
@@ -11,6 +11,8 @@ import com.liferay.portal.kernel.model.MVCCModel;
 import com.liferay.portal.kernel.model.ShardedModel;
 import com.liferay.portal.kernel.model.StagedAuditedModel;
 
+import java.math.BigDecimal;
+
 import java.util.Date;
 
 import org.osgi.annotation.versioning.ProviderType;
@@ -231,28 +233,28 @@ public interface CommerceInventoryWarehouseItemModel
 	 *
 	 * @return the quantity of this commerce inventory warehouse item
 	 */
-	public int getQuantity();
+	public BigDecimal getQuantity();
 
 	/**
 	 * Sets the quantity of this commerce inventory warehouse item.
 	 *
 	 * @param quantity the quantity of this commerce inventory warehouse item
 	 */
-	public void setQuantity(int quantity);
+	public void setQuantity(BigDecimal quantity);
 
 	/**
 	 * Returns the reserved quantity of this commerce inventory warehouse item.
 	 *
 	 * @return the reserved quantity of this commerce inventory warehouse item
 	 */
-	public int getReservedQuantity();
+	public BigDecimal getReservedQuantity();
 
 	/**
 	 * Sets the reserved quantity of this commerce inventory warehouse item.
 	 *
 	 * @param reservedQuantity the reserved quantity of this commerce inventory warehouse item
 	 */
-	public void setReservedQuantity(int reservedQuantity);
+	public void setReservedQuantity(BigDecimal reservedQuantity);
 
 	/**
 	 * Returns the sku of this commerce inventory warehouse item.
